Revert "making corrections to routes scafolding and deleting payment.js file …"
