+++ src/routes/userRouter.js
@@ -1,83 +1,50 @@
 const { Router } = require("express");
-const {
-  create,
-  authenticate,
-  findById,
-  pets,
-  subscription,
-  pickups,
-} = require("../../usecases/user");
-const { authHandler } = require("../../middlewares/authHandler");
+const { create, authenticate, findById, pets, subscription, pickups } = require("../usecases/user");
+const { authHandler } = require("../middlewares/authHandler");
 
 const routes = Router();
 
 routes.post("/", async (req, res) => {
-  const {
-    email,
-    password,
-    firstName,
-    lastName,
-    phone,
-    street,
-    number,
-    interior,
-    neighborhood,
-    municipality,
-    state,
-    zipCode,
-  } = req.body;
+    const {email, password, firstName, lastName, phone, street, number, interior, neighborhood, municipality, state, zipCode} = req.body;
 
-  try {
-    const payload = await create({
-      email,
-      password,
-      firstName,
-      lastName,
-      phone,
-      street,
-      number,
-      interior,
-      neighborhood,
-      municipality,
-      state,
-      zipCode,
-    });
-    res.json({ ok: true, payload });
-  } catch (error) {
-    const { message } = error;
-    res.status(400).json({ ok: false, message });
-  }
+    try {
+        const payload = await create({email, password, firstName, lastName, phone, street, number, interior, neighborhood, municipality, state, zipCode});
+        res.json({ ok: true, payload });
+    } catch (error) {
+        const { message } = error;
+        res.status(400).json({ ok: false, message });
+    }
 });
 
 routes.get("/", authHandler, async (req, res) => {
-  const userId = req.params.token.sub;
-  try {
-    const payload = await findById(userId);
-    res.status(202).json({ ok: true, payload });
-  } catch (error) {
-    const { message } = error;
-    res.status(401).json({ ok: false, message });
-  }
+    const userId = req.params.token.sub;
+    try {
+      const payload = await findById(userId);
+      res.status(202).json({ ok: true, payload });
+    } catch (error) {
+      const  { message } = error;
+      res.status(401).json({ ok: false, message });
+    }
 });
 
 routes.post("/auth", async (req, res) => {
-  const { email, password } = req.body;
-
-  try {
-    const payload = await authenticate(email, password);
-    res.status(202).json({ ok: true, payload });
-  } catch (error) {
-    const { message } = error;
-    res.status(401).json({ ok: false, message });
-  }
-});
+    const { email, password } = req.body;
+  
+    try {
+      const payload = await authenticate(email, password);
+      res.status(202).json({ ok: true, payload });
+    } catch (error) {
+      const { message } = error;
+      res.status(401).json({ ok: false, message });
+    }
+  });
 
 routes.post("/pet", authHandler, async (req, res) => {
   const userId = req.params.token.sub;
   const { name, size, species } = req.body;
 
   try {
-    const payload = await pets.addPet({ userId, name, size, species });
+    const payload = await pets.addPet({userId, name, size, species});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -90,7 +57,7 @@ routes.delete("/pet", authHandler, async (req, res) => {
   const { name } = req.body;
 
   try {
-    const payload = await pets.delPet({ userId, name });
+    const payload = await pets.delPet({userId, name});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -102,7 +69,7 @@ routes.get("/allPets", authHandler, async (req, res) => {
   const userId = req.params.token.sub;
 
   try {
-    const payload = await pets.getAllPets({ userId });
+    const payload = await pets.getAllPets({userId});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -114,7 +81,7 @@ routes.get("/totalFee", authHandler, async (req, res) => {
   const userId = req.params.token.sub;
 
   try {
-    const payload = await subscription.calcTotalFee({ userId });
+    const payload = await subscription.calcTotalFee({userId});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -126,7 +93,7 @@ routes.get("/initialFee", authHandler, async (req, res) => {
   const userId = req.params.token.sub;
 
   try {
-    const payload = await subscription.calcInitialFee({ userId });
+    const payload = await subscription.calcInitialFee({userId});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -139,10 +106,7 @@ routes.post("/package", authHandler, async (req, res) => {
   const { packageId } = req.body;
 
   try {
-    const payload = await subscription.addPackage({
-      userId,
-      package: packageId,
-    });
+    const payload = await subscription.addPackage({userId, package: packageId});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -155,10 +119,7 @@ routes.delete("/package", authHandler, async (req, res) => {
   const { packageId } = req.body;
 
   try {
-    const payload = await subscription.removePackage({
-      userId,
-      package: packageId,
-    });
+    const payload = await subscription.removePackage({userId, package: packageId});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -170,7 +131,7 @@ routes.get("/allPackages", authHandler, async (req, res) => {
   const userId = req.params.token.sub;
 
   try {
-    const payload = await subscription.getAllPackages({ userId });
+    const payload = await subscription.getAllPackages({userId});
     res.status(202).json({ ok: true, payload });
   } catch (error) {
     const { message } = error;
@@ -180,7 +141,7 @@ routes.get("/allPackages", authHandler, async (req, res) => {
 
 routes.get("/nextPickup", authHandler, async (req, res) => {
   const userId = req.params.token.sub;
-
+  
   try {
     const payload = await pickups.getNextPickup(userId);
     res.status(202).json({ ok: true, payload });
